Categorias

+++ TrabalhoPO/TrabalhoPO/Controllers/ProdutosController.cs
@@ -12,6 +12,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Index()
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             ViewBag.Categorias = db.Categorias.ToList();
 
             return View(db.Produtos.ToList());
@@ -19,6 +24,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Detalhes(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             ViewBag.Categorias = db.Categorias.ToList();
 
             return View(db.Produtos.Find(id));
@@ -26,6 +36,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Editar(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             ViewBag.Categorias = db.Categorias.ToList();
 
             return View(db.Produtos.Find(id));
@@ -33,6 +48,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Criar()
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             ViewBag.Categorias = db.Categorias.ToList();
 
             return View(new Produto());
@@ -40,6 +60,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Excluir(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             Produto produto = db.Produtos.Find(id);
             db.Produtos.Remove(produto);
             db.SaveChanges();
@@ -49,12 +74,22 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult ExcluirModal(int id)
         {
-            return PartialView(db.Produtos.Find(id));
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
+            return PartialView("~/Views/Shared/_ModalExcluir.cshtml", db.Produtos.Find(id));
         }
 
         [HttpPost]
         public ActionResult Salvar(Produto produto)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             produto.AtualizaCampos();
             db.Set<Produto>().AddOrUpdate(produto);
             db.SaveChanges();
@@ -64,6 +99,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Insere(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             Produto produto = db.Produtos.Find(id);
             produto.Insere(1);
             db.SaveChanges();
@@ -73,6 +113,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult Retira(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             Produto produto = db.Produtos.Find(id);
             produto.Retira(1);
             db.SaveChanges();
@@ -82,6 +127,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult AumentaMinimo(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             Produto produto = db.Produtos.Find(id);
             produto.SetEstoqueMinimo(produto.EstoqueMinimo + 1);
             db.SaveChanges();
@@ -91,6 +141,11 @@ namespace TrabalhoPO.Controllers
 
         public ActionResult DiminuiMinimo(int id)
         {
+            if (Session != null && Session["usuarioLogadoID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
+
             Produto produto = db.Produtos.Find(id);
             produto.SetEstoqueMinimo(produto.EstoqueMinimo - 1);
             db.SaveChanges();
